backend/requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
SQLAlchemy==2.0.36
PyMySQL==1.1.1
numpy==2.2.4
Flask-Cors==4.0.1
Flask-MySQLdb==2.0.0
mysqlclient==2.2.7